pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rita"
version = "0.1.0"
description = "Render it then argue, a complementary helm toolkit for validation and template creation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [
    { name = "sam.mckay.loureiro", email = "sam@bluebicycle.tv" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
keywords = []
dependencies = [
    "boto3>=1.42.50",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich-click>=1.9.7",
    "typer>=0.12.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "ty>=0.0.14",
    "hatch>=1.16.3",
    "prek>=0.1.0",
    "pysentry-rs>=0.1.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.7.0",
    "mkdocstrings-python>=2.0.1",
]

[project.urls]
Homepage = "https://github.com/SMLoureiro/rita"
Documentation = "https://SMLoureiro.github.io/rita"
Repository = "https://github.com/SMLoureiro/rita"
Issues = "https://github.com/SMLoureiro/rita/issues"
Changelog = "https://github.com/SMLoureiro/rita/blob/main/CHANGELOG.md"

[project.scripts]
rita = "rita.cli:main"

# =============================================================================
# Hatch
# =============================================================================

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rita"]

# Matrix testing across Python versions
[tool.hatch.envs.test]

[[tool.hatch.envs.test.matrix]]
python = ["3.10", "3.11", "3.12", "3.13"]

# =============================================================================
# Ruff
# =============================================================================

[tool.ruff]
target-version = "py313"
line-length = 110
src = ["src", "tests"]

[tool.ruff.lint]
ignore = ["E501"]  # Line too long - handled by formatter
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "RUF",    # Ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["rita"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001"]

# =============================================================================
# ty (Astral type checker)
# =============================================================================

[tool.ty.src]
include = ["**/*.py"]

# =============================================================================
# Pytest
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

# =============================================================================
# Coverage
# =============================================================================

[tool.coverage.run]
source = ["src/rita"]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
show_missing = true

# =============================================================================
# git-cliff (Changelog generation)
# =============================================================================

[tool.git-cliff]
config = "cliff.toml"
